CPU Batched chol local energy estimate
Batching occurs by forming a stack of walker G-functions to combine
smaller gemms in to one bigger gem and avoid a loop.  At the expense of
a reshape we can evaluate exchange component with bigger gemm and a
tensor contraction.  In the jax implementation we will vmap over chol
index.