Update actions checkout and setup-python (fixes warning). (#277)
The mpi tests failing was a red herring. The examples were failing
because of #278, this sent a signal to the other tests to fail but I'm
guessing the error was only received on the root process so the MPI jobs
would hang. I've set a timeout for these for the moment but there may be
a more sensible thing to do.