Remove dot_real_cmplx function from wicks kernels. (#276)
This function assumed the cholesky intermediate was real in an attempt
to save some multiplications. This assumption is incorrect in general.
This fortunately did not lead to incorrect results as the formula held
for both real and complex input A, which explains why the unit tests
passed. Removes this in favour of * for element wise multiplication.
Also force more complex walkers in the tests to exercise the code path.

Fixes #275