Fix wicks assertion (#287)
Fixes #285. This wasn't a bug per se, more of a case of poor assertions.
The optimized variational wavefunction code uses some number (DET_LEN)
64 bit unsigned integers to represent the bitstrings. This was hardcoded
to 2 so, the max number of spin orbitals would be 128. If one went
beyond this the code wouldn't crash. Here I extend this to 256 and also
add an assertion when build the trial wavefunction class. I also added
some better unit tests and test the wicks_helper as part of the CI.

This code is hopefully going to go away if I ever get the time.